requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
faker==26.0.0
fastapi==0.115.5
uvicorn==0.30.6
fastapi==0.115.0
pydantic==2.9.0
streamlit==1.38.0